Spring boot security and Hardcoded Data

+++ src/main/java/hua/WebApp/SpringBoot/Controller.java
@@ -4,71 +4,134 @@ import hua.WebApp.SpringBoot.entities.RecommendationLetter.RecommendationLetter;
 import hua.WebApp.SpringBoot.entities.RecommendationLetter.RecommendationLetterRepository;
 import hua.WebApp.SpringBoot.entities.Request.Request;
 import hua.WebApp.SpringBoot.entities.Request.RequestRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.List;
+import java.util.Optional;
 
 
 @RestController
 public class Controller {
 
+    @Autowired
+    UserDetailsService userDetailsService;
 
     private final RequestRepository requestRepository;
     private final RecommendationLetterRepository recommendationLetterRepository;
 
-    public Controller(RequestRepository requestRepository, RecommendationLetterRepository recommendationLetterRepository){
+
+
+    public Controller(RequestRepository requestRepository, RecommendationLetterRepository recommendationLetterRepository, UserRepository userRepository){
         this.requestRepository = requestRepository;
         this.recommendationLetterRepository = recommendationLetterRepository;
+
     }
 
     //Student Endpoints
 
+
     @GetMapping("/requests")
-    public List<Request> getAllStudents(){
-        return requestRepository.findAll();
+    public List<Request> getAllRequests(){
+
+
+        List<Request> requests = requestRepository.findAll();
+        List<Request> userRequests = null;
+       for (Request r:requests) {
+           if(r.getUid().equals(GetUsername()))
+               userRequests.add(r);
+
+       }
+
+        return userRequests;
     }
 
     @PostMapping("/addRequest")
     public String addRequest (@RequestBody Request r){
+       r.setUid(GetUsername());
         requestRepository.save(r);
         return "Request Saved";
     }
     @PutMapping("/editRequest/{requestId}")
     public String editRequest(@PathVariable("requestId") Long requestId){
-       // requestRepository.save(requestId);
+       //TODO
 
         return "Request Updated";
     }
 
     @DeleteMapping( "deleteRequest/{requestId}")
-    public void deleteRequest(@PathVariable("requestId") Long requestId){
-        requestRepository.findById(requestId);
+    public String deleteRequest(@PathVariable("requestId") Long requestId){
+        Optional<Request> r = requestRepository.findById(requestId);
 
         boolean exists =  requestRepository.existsById(requestId);
-        if (!exists){
+
+        if (!exists ){
             throw  new IllegalStateException(
-                    "student with id : "+ requestId + " does not exists"
+                    "request with id : "+ requestId + " does not exists"
+            );
+        }
+        if (!r.get().getUid().equals(GetUsername()) ){
+            throw  new IllegalStateException(
+                    "You dont have access to delete the request with id: "+ requestId
             );
         }
         requestRepository.deleteById(requestId);
+        return "Request Deleted";
     }
 
     //Professor Endpoints
-    @PutMapping("/setRequestStatus")
-    public String setRequestStatus(@RequestBody Request r){
-        requestRepository.save(r);
-        return "Request Updated";
+    @GetMapping("/pendingRequests")
+    public List<Request> getAllPendingRequests(){
+
+
+        List<Request> requests = requestRepository.findAll();
+        List<Request> pendingRequests = null;
+        for (Request r:requests) {
+            if(r.getDest().equals(GetUsername()))
+                pendingRequests.add(r);
+
+        }
+
+        return pendingRequests;
+    }
+
+    @PutMapping("/setRequestStatus/{requestId}")
+    public String setRequestStatus(@PathVariable(value = "id") Long requestId,
+                                   @Valid @RequestBody Request requestDetails){
+
+        boolean exists =  requestRepository.existsById(requestId);
+        if (!exists ){
+            throw  new IllegalStateException(
+                    "request with id : "+ requestId + " does not exists"
+            );
+        }
+        Optional<Request> r =  requestRepository.findById(requestId);
+        if (!r.get().getDest().equals(GetUsername()) ){
+            throw  new IllegalStateException(
+                    "You dont have access to change the status of this request with id: "+ requestId
+            );
+        }
+
+        Request request =  r.get();
+        request.setAgreed(requestDetails.isAgreed());
+        requestRepository.save(request);
+
+        return  "Request Updated";
     }
 
     @PostMapping("/addLetter")
     public String addRecommendationLetter (@RequestBody RecommendationLetter rl){
+        rl.setUid(GetUsername());
         recommendationLetterRepository.save(rl);
         return "Recommendation Letter Saved";
     }
     @PutMapping("/editLetter/{letterId}")
     public String editLetter(@PathVariable("letterId") Long letterId){
-        // RecommendationLetterRepository.save(letterId);
-
+       //TODO
         return "Recommendation Letter Updated";
     }
     @DeleteMapping( "deleteLetter/{letterId}")
@@ -81,7 +144,26 @@ public class Controller {
                     "student with id : "+ letterId + " does not exists"
             );
         }
+        Optional<RecommendationLetter> recommendationLetter = recommendationLetterRepository.findById(letterId);
+        if (!recommendationLetter.get().getUid().equals(GetUsername()) ){
+            throw  new IllegalStateException(
+                    "You dont have access to change the status of this request with id: "+ letterId
+            );
+        }
         recommendationLetterRepository.deleteById(letterId);
+    }
+
+    private String GetUsername(){
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String username;
+        if (principal instanceof UserDetails) {
+            username = ((UserDetails)principal).getUsername();
+
+        } else {
+            username = principal.toString();
+
+        }
+        return username;
     }
